% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_etas.R
\name{fetch_etas}
\alias{fetch_etas}
\title{Fetches the etas from an lst object}
\usage{
fetch_etas(lst, rse_digits = NA)
}
\arguments{
\item{lst}{A list object containing the output from a NONMEM run}

\item{rse_digits}{The number of digits to round the RSEs to}
}
\value{
A tibble containing the etas and their RSEs
}
\description{
Fetches the etas from an lst object
}
\examples{
fetch_etas(lst = lst)
}
