% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sigmas.R
\name{fetch_sigmas}
\alias{fetch_sigmas}
\title{Fetches the sigmas from an lst object}
\usage{
fetch_sigmas(lst, rse_digits = NA)
}
\arguments{
\item{lst}{A list object containing the output from a NONMEM run}

\item{rse_digits}{The number of digits to round the RSEs to}

\item{shk_digits}{The number of digits to round the SHKs to}
}
\value{
A tibble containing the sigmas and their RSEs
}
\description{
Fetches the sigmas from an lst object
}
\examples{
fetch_sigmas(lst = lst)
}
