% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_thetas.R
\name{fetch_thetas}
\alias{fetch_thetas}
\title{Fetch Thetas from List}
\usage{
fetch_thetas(lst, rse_digits = NA)
}
\arguments{
\item{lst}{A lst-file object.}

\item{rse_digits}{An optional parameter specifying the number of digits for rounding standard errors. Default is NA.}
}
\value{
A data frame containing the extracted theta values and standard errors.
}
\description{
This function extracts theta values from a lst-file object.
}
